refactor: userConfigの保存処理を分離
