+++ main/src/pages/tests/test_page/userConfigContext.tsx
@@ -7,15 +7,10 @@ import {
   type YearConfigEntry,
 } from "./initialUserConfig";
 import { type UserConfigContextValue, UserConfigContext } from "./userConfigStore";
+import { loadUserConfig, persistUserConfig } from "./userConfigStorage";
 import { PLAYER_REGISTRY_UPDATED_EVENT } from "@/data/userYearRegistry";
 import type { YearKey } from "@/data/vocabLoader";
 import { getAllRegistry } from "@/hooks/getAllRegistry";
-
-const USER_CONFIG_STORAGE_KEY = "user-config:max-count";
-
-const isRecord = (value: unknown): value is Record<string, unknown> => {
-  return typeof value === "object" && value !== null;
-};
 
 // レジストリの最新状態に合わせてyear設定を補完/整理する
 const syncConfigWithRegistry = (current: UserConfigState): UserConfigState => {
@@ -48,47 +43,13 @@ const syncConfigWithRegistry = (current: UserConfigState): UserConfigState => {
   };
 };
 
-const loadStoredConfig = (): UserConfigState => {
-  if (typeof window === "undefined") return initialUserConfig;
-  try {
-    const raw = window.localStorage.getItem(USER_CONFIG_STORAGE_KEY);
-    if (!raw) return initialUserConfig;
-    const parsed = JSON.parse(raw) as unknown;
-    if (!isRecord(parsed)) return initialUserConfig;
-
-    // 旧形式（年度ごとの設定だけ）も吸収して互換性を保つ
-    if (!("years" in parsed)) {
-      const legacy = parsed as Record<string, YearConfigEntry>;
-      return syncConfigWithRegistry({
-        ...initialUserConfig,
-        years: {
-          ...initialUserConfig.years,
-          ...legacy,
-        },
-      });
-    }
-
-    const nextSoundPreference: SoundPreferenceState = {
-      ...initialUserConfig.soundPreference,
-      ...(isRecord(parsed.soundPreference) ? parsed.soundPreference : {}),
-    };
-    const nextYears = isRecord(parsed.years)
-      ? { ...initialUserConfig.years, ...(parsed.years as Record<string, YearConfigEntry>) }
-      : initialUserConfig.years;
-    return syncConfigWithRegistry({
-      ...initialUserConfig,
-      ...parsed,
-      years: nextYears,
-      soundPreference: nextSoundPreference,
-    });
-  } catch (error) {
-    console.warn("Failed to load user config", error);
-    return initialUserConfig;
-  }
-};
-
 export function UserConfigProvider({ children }: { children: ReactNode }) {
-  const [config, setConfig] = useState<UserConfigState>(() => loadStoredConfig());
+  const [config, setConfig] = useState<UserConfigState>(() =>
+    loadUserConfig({
+      initialConfig: initialUserConfig,
+      syncConfigWithRegistry,
+    })
+  );
 
   useEffect(() => {
     const syncConfig = () => {
@@ -106,12 +67,7 @@ export function UserConfigProvider({ children }: { children: ReactNode }) {
   }, []);
 
   useEffect(() => {
-    if (typeof window === "undefined") return;
-    try {
-      window.localStorage.setItem(USER_CONFIG_STORAGE_KEY, JSON.stringify(config));
-    } catch (error) {
-      console.warn("Failed to persist user config", error);
-    }
+    persistUserConfig(config);
   }, [config]);
 
   const setMaxCount = useCallback((year: YearKey, value: number) => {
